PHP 8.5 deprecations

+++ npgCore/extensions/elFinder/php/elFinder.class.php
@@ -4297,7 +4297,7 @@ var go = function() {
 
 	protected function utime() {
 		$time = explode(" ", microtime());
-		return (double) $time[1] + (double) $time[0];
+		return (float) $time[1] + (float) $time[0];
 	}
 
 	/**
